-added blending mode to the render api -adjust the font service api to return also the GlyphMetric information -added support for window position and size in the the form class -adjust the Settings class to use String instead of string literal -extend the Server class to send a message to all sockets -the opengl renderer use material objects instead of a place holder id -fixed a wrong type usage in the Windows GDI event handling implementation

+++ services/window/GDI/WindowsWindow.cpp
@@ -156,8 +156,8 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
     }
     case WM_MOVE:
         wnd = dynamic_cast<WindowsWindow*>(WindowsWindow::GetObjectByHandle(hWnd));
-        pos.X = LOWORD(lParam);
-        pos.Y = HIWORD(lParam);
+        pos.X = (int)(short)LOWORD(lParam);
+        pos.Y = (int)(short)HIWORD(lParam);
         wnd->OnReposition(pos);
         return 0;
     case WM_SIZE:
